+++ src/polymer/pseudo-element-scanner.ts
@@ -17,10 +17,10 @@ import {ASTNode} from 'parse5';
 
 import {HtmlVisitor, ParsedHtmlDocument} from '../html/html-document';
 import {HtmlScanner} from '../html/html-scanner';
+import {getOrInferPrivacy} from '../javascript/esutil';
 import * as jsdoc from '../javascript/jsdoc';
 
 import {annotateElementHeader} from './docs';
-import {getOrInferPrivacy} from './js-utils';
 import {ScannedPolymerElement} from './polymer-element';
 
 /**
@@ -58,6 +58,7 @@ export class PseudoElementScanner implements HtmlScanner {
             behaviors: [],
             className: undefined,
             extends: undefined,
+            staticMethods: new Map(),
             methods: new Map(),
             mixins: [],
             observers: [],